Base error class and children of it now work
I will still look at it to see if I can make it better. I'm not
that used to with Java so need to study if I should use interfaces
or use inheritance (or do something else).

This works for now. I'm overriding the getters so they refer to
the variable present in each child class and not refer to
uninitialized variables in the base class.